added sounds and worked on linear progression
